+++ test/unit/internal/oidc/test_ambient.py
@@ -16,7 +16,6 @@ import json
 
 import pretend
 import pytest
-from requests import HTTPError, Timeout
 
 from id import detect_credential
 from id._internal.oidc import ambient
@@ -93,22 +92,22 @@ def test_detect_github_request_fails(monkeypatch):
     monkeypatch.setenv("ACTIONS_ID_TOKEN_REQUEST_URL", "fakeurl")
 
     resp = pretend.stub(
-        raise_for_status=pretend.raiser(HTTPError),
-        status_code=999,
-        content=b"something",
+        status=999,
+        data=b"something",
     )
-    requests = pretend.stub(get=pretend.call_recorder(lambda url, **kw: resp), HTTPError=HTTPError)
-    monkeypatch.setattr(ambient, "requests", requests)
+    u3 = pretend.stub(request=pretend.call_recorder(lambda meth, url, **kw: resp))
+    monkeypatch.setattr(ambient, "urllib3", u3)
 
     with pytest.raises(
         ambient.AmbientCredentialError,
         match=r"GitHub: OIDC token request failed \(code=999, body='something'\)",
     ):
         ambient.detect_github("some-audience")
-    assert requests.get.calls == [
+    assert u3.request.calls == [
         pretend.call(
+            "GET",
             "fakeurl",
-            params={"audience": "some-audience"},
+            fields={"audience": "some-audience"},
             headers={"Authorization": "bearer faketoken"},
             timeout=30,
         )
@@ -120,27 +119,17 @@ def test_detect_github_request_timeout(monkeypatch):
     monkeypatch.setenv("ACTIONS_ID_TOKEN_REQUEST_TOKEN", "faketoken")
     monkeypatch.setenv("ACTIONS_ID_TOKEN_REQUEST_URL", "fakeurl")
 
-    resp = pretend.stub(raise_for_status=pretend.raiser(Timeout))
-    requests = pretend.stub(
-        get=pretend.call_recorder(lambda url, **kw: resp),
-        HTTPError=HTTPError,
-        Timeout=Timeout,
+    u3 = pretend.stub(
+        request=pretend.raiser(ValueError), exceptions=pretend.stub(MaxRetryError=ValueError)
     )
-    monkeypatch.setattr(ambient, "requests", requests)
+
+    monkeypatch.setattr(ambient, "urllib3", u3)
 
     with pytest.raises(
         ambient.AmbientCredentialError,
         match=r"GitHub: OIDC token request timed out",
     ):
         ambient.detect_github("some-audience")
-    assert requests.get.calls == [
-        pretend.call(
-            "fakeurl",
-            params={"audience": "some-audience"},
-            headers={"Authorization": "bearer faketoken"},
-            timeout=30,
-        )
-    ]
 
 
 def test_detect_github_invalid_json_payload(monkeypatch):
@@ -148,19 +137,20 @@ def test_detect_github_invalid_json_payload(monkeypatch):
     monkeypatch.setenv("ACTIONS_ID_TOKEN_REQUEST_TOKEN", "faketoken")
     monkeypatch.setenv("ACTIONS_ID_TOKEN_REQUEST_URL", "fakeurl")
 
-    resp = pretend.stub(raise_for_status=lambda: None, json=pretend.raiser(json.JSONDecodeError))
-    requests = pretend.stub(get=pretend.call_recorder(lambda url, **kw: resp))
-    monkeypatch.setattr(ambient, "requests", requests)
+    resp = pretend.stub(status=200, json=pretend.raiser(json.JSONDecodeError))
+    request = pretend.call_recorder(lambda meth, url, **kw: resp)
+    monkeypatch.setattr(ambient.urllib3, "request", request)
 
     with pytest.raises(
         ambient.AmbientCredentialError,
         match="GitHub: malformed or incomplete JSON",
     ):
         ambient.detect_github("some-audience")
-    assert requests.get.calls == [
+    assert request.calls == [
         pretend.call(
+            "GET",
             "fakeurl",
-            params={"audience": "some-audience"},
+            fields={"audience": "some-audience"},
             headers={"Authorization": "bearer faketoken"},
             timeout=30,
         )
@@ -173,19 +163,20 @@ def test_detect_github_bad_payload(monkeypatch, payload):
     monkeypatch.setenv("ACTIONS_ID_TOKEN_REQUEST_TOKEN", "faketoken")
     monkeypatch.setenv("ACTIONS_ID_TOKEN_REQUEST_URL", "fakeurl")
 
-    resp = pretend.stub(raise_for_status=lambda: None, json=pretend.call_recorder(lambda: payload))
-    requests = pretend.stub(get=pretend.call_recorder(lambda url, **kw: resp))
-    monkeypatch.setattr(ambient, "requests", requests)
+    resp = pretend.stub(status=200, json=pretend.call_recorder(lambda: payload))
+    u3 = pretend.stub(request=pretend.call_recorder(lambda meth, url, **kw: resp))
+    monkeypatch.setattr(ambient, "urllib3", u3)
 
     with pytest.raises(
         ambient.AmbientCredentialError,
         match="GitHub: malformed or incomplete JSON",
     ):
         ambient.detect_github("some-audience")
-    assert requests.get.calls == [
+    assert u3.request.calls == [
         pretend.call(
+            "GET",
             "fakeurl",
-            params={"audience": "some-audience"},
+            fields={"audience": "some-audience"},
             headers={"Authorization": "bearer faketoken"},
             timeout=30,
         )
@@ -199,17 +190,18 @@ def test_detect_github(monkeypatch):
     monkeypatch.setenv("ACTIONS_ID_TOKEN_REQUEST_URL", "fakeurl")
 
     resp = pretend.stub(
-        raise_for_status=lambda: None,
+        status=200,
         json=pretend.call_recorder(lambda: {"value": "fakejwt"}),
     )
-    requests = pretend.stub(get=pretend.call_recorder(lambda url, **kw: resp))
-    monkeypatch.setattr(ambient, "requests", requests)
+    u3 = pretend.stub(request=pretend.call_recorder(lambda meth, url, **kw: resp))
+    monkeypatch.setattr(ambient, "urllib3", u3)
 
     assert ambient.detect_github("some-audience") == "fakejwt"
-    assert requests.get.calls == [
+    assert u3.request.calls == [
         pretend.call(
+            "GET",
             "fakeurl",
-            params={"audience": "some-audience"},
+            fields={"audience": "some-audience"},
             headers={"Authorization": "bearer faketoken"},
             timeout=30,
         )
@@ -223,13 +215,9 @@ def test_gcp_impersonation_access_token_request_fail(monkeypatch):
     logger = pretend.stub(debug=pretend.call_recorder(lambda s: None))
     monkeypatch.setattr(ambient, "logger", logger)
 
-    resp = pretend.stub(
-        raise_for_status=pretend.raiser(HTTPError),
-        status_code=999,
-        content=b"something",
-    )
-    requests = pretend.stub(get=pretend.call_recorder(lambda url, **kw: resp), HTTPError=HTTPError)
-    monkeypatch.setattr(ambient, "requests", requests)
+    resp = pretend.stub(status=999, data=b"something")
+    u3 = pretend.stub(request=pretend.call_recorder(lambda meth, url, **kw: resp))
+    monkeypatch.setattr(ambient, "urllib3", u3)
 
     with pytest.raises(
         ambient.AmbientCredentialError,
@@ -250,13 +238,11 @@ def test_gcp_impersonation_access_token_request_timeout(monkeypatch):
     logger = pretend.stub(debug=pretend.call_recorder(lambda s: None))
     monkeypatch.setattr(ambient, "logger", logger)
 
-    resp = pretend.stub(raise_for_status=pretend.raiser(Timeout))
-    requests = pretend.stub(
-        get=pretend.call_recorder(lambda url, **kw: resp),
-        HTTPError=HTTPError,
-        Timeout=Timeout,
+    u3 = pretend.stub(
+        request=pretend.raiser(ValueError), exceptions=pretend.stub(MaxRetryError=ValueError)
     )
-    monkeypatch.setattr(ambient, "requests", requests)
+
+    monkeypatch.setattr(ambient, "urllib3", u3)
 
     with pytest.raises(
         ambient.AmbientCredentialError,
@@ -277,9 +263,9 @@ def test_gcp_impersonation_access_token_missing(monkeypatch):
     logger = pretend.stub(debug=pretend.call_recorder(lambda s: None))
     monkeypatch.setattr(ambient, "logger", logger)
 
-    resp = pretend.stub(raise_for_status=lambda: None, json=lambda: {})
-    requests = pretend.stub(get=pretend.call_recorder(lambda url, **kw: resp))
-    monkeypatch.setattr(ambient, "requests", requests)
+    resp = pretend.stub(status=200, json=lambda: {})
+    u3 = pretend.stub(request=pretend.call_recorder(lambda meth, url, **kw: resp))
+    monkeypatch.setattr(ambient, "urllib3", u3)
 
     with pytest.raises(
         ambient.AmbientCredentialError,
@@ -301,20 +287,22 @@ def test_gcp_impersonation_identity_token_request_fail(monkeypatch):
     monkeypatch.setattr(ambient, "logger", logger)
 
     access_token = pretend.stub()
-    get_resp = pretend.stub(
-        raise_for_status=lambda: None, json=lambda: {"access_token": access_token}
-    )
+    get_resp = pretend.stub(status=200, json=lambda: {"access_token": access_token})
     post_resp = pretend.stub(
-        raise_for_status=pretend.raiser(HTTPError),
-        status_code=999,
-        content=b"something",
+        status=999,
+        data=b"something",
     )
-    requests = pretend.stub(
-        get=pretend.call_recorder(lambda url, **kw: get_resp),
-        post=pretend.call_recorder(lambda url, **kw: post_resp),
-        HTTPError=HTTPError,
-    )
-    monkeypatch.setattr(ambient, "requests", requests)
+
+    def _request(meth, *a, **kw):
+        if meth == "GET":
+            return get_resp
+        elif meth == "POST":
+            return post_resp
+        else:
+            assert False
+
+    u3 = pretend.stub(request=_request)
+    monkeypatch.setattr(ambient, "urllib3", u3)
 
     with pytest.raises(
         ambient.AmbientCredentialError,
@@ -337,17 +325,18 @@ def test_gcp_impersonation_identity_token_request_timeout(monkeypatch):
     monkeypatch.setattr(ambient, "logger", logger)
 
     access_token = pretend.stub()
-    get_resp = pretend.stub(
-        raise_for_status=lambda: None, json=lambda: {"access_token": access_token}
-    )
-    post_resp = pretend.stub(raise_for_status=pretend.raiser(Timeout))
-    requests = pretend.stub(
-        get=pretend.call_recorder(lambda url, **kw: get_resp),
-        post=pretend.call_recorder(lambda url, **kw: post_resp),
-        HTTPError=HTTPError,
-        Timeout=Timeout,
-    )
-    monkeypatch.setattr(ambient, "requests", requests)
+    get_resp = pretend.stub(status=200, json=lambda: {"access_token": access_token})
+
+    def _request(meth, *a, **kw):
+        if meth == "GET":
+            return get_resp
+        elif meth == "POST":
+            raise ValueError
+        else:
+            assert False
+
+    u3 = pretend.stub(request=_request, exceptions=pretend.stub(MaxRetryError=ValueError))
+    monkeypatch.setattr(ambient, "urllib3", u3)
 
     with pytest.raises(
         ambient.AmbientCredentialError,
@@ -370,16 +359,19 @@ def test_gcp_impersonation_identity_token_missing(monkeypatch):
     monkeypatch.setattr(ambient, "logger", logger)
 
     access_token = pretend.stub()
-    get_resp = pretend.stub(
-        raise_for_status=lambda: None, json=lambda: {"access_token": access_token}
-    )
-    post_resp = pretend.stub(raise_for_status=lambda: None, json=lambda: {})
-    requests = pretend.stub(
-        get=pretend.call_recorder(lambda url, **kw: get_resp),
-        post=pretend.call_recorder(lambda url, **kw: post_resp),
-        HTTPError=HTTPError,
-    )
-    monkeypatch.setattr(ambient, "requests", requests)
+    get_resp = pretend.stub(status=200, json=lambda: {"access_token": access_token})
+    post_resp = pretend.stub(status=200, json=lambda: {})
+
+    def _request(meth, *a, **kw):
+        if meth == "GET":
+            return get_resp
+        elif meth == "POST":
+            return post_resp
+        else:
+            assert False
+
+    u3 = pretend.stub(request=_request)
+    monkeypatch.setattr(ambient, "urllib3", u3)
 
     with pytest.raises(
         ambient.AmbientCredentialError,
@@ -403,16 +395,19 @@ def test_gcp_impersonation_succeeds(monkeypatch):
 
     access_token = pretend.stub()
     oidc_token = pretend.stub()
-    get_resp = pretend.stub(
-        raise_for_status=lambda: None, json=lambda: {"access_token": access_token}
-    )
-    post_resp = pretend.stub(raise_for_status=lambda: None, json=lambda: {"token": oidc_token})
-    requests = pretend.stub(
-        get=pretend.call_recorder(lambda url, **kw: get_resp),
-        post=pretend.call_recorder(lambda url, **kw: post_resp),
-        HTTPError=HTTPError,
-    )
-    monkeypatch.setattr(ambient, "requests", requests)
+    get_resp = pretend.stub(status=200, json=lambda: {"access_token": access_token})
+    post_resp = pretend.stub(status=200, json=lambda: {"token": oidc_token})
+
+    def _request(meth, *a, **kw):
+        if meth == "GET":
+            return get_resp
+        elif meth == "POST":
+            return post_resp
+        else:
+            assert False
+
+    u3 = pretend.stub(request=_request)
+    monkeypatch.setattr(ambient, "urllib3", u3)
 
     assert ambient.detect_gcp("some-audience") == oidc_token
 
@@ -469,22 +464,22 @@ def test_detect_gcp_request_fails(monkeypatch):
     monkeypatch.setattr(ambient, "_open", lambda fn: stub_file)  # type: ignore
 
     resp = pretend.stub(
-        raise_for_status=pretend.raiser(HTTPError),
-        status_code=999,
-        content=b"something",
+        status=999,
+        data=b"something",
     )
-    requests = pretend.stub(get=pretend.call_recorder(lambda url, **kw: resp), HTTPError=HTTPError)
-    monkeypatch.setattr(ambient, "requests", requests)
+    u3 = pretend.stub(request=pretend.call_recorder(lambda meth, url, **kw: resp))
+    monkeypatch.setattr(ambient, "urllib3", u3)
 
     with pytest.raises(
         ambient.AmbientCredentialError,
         match=r"GCP: OIDC token request failed \(code=999, body='something'\)",
     ):
         ambient.detect_gcp("some-audience")
-    assert requests.get.calls == [
+    assert u3.request.calls == [
         pretend.call(
+            "GET",
             ambient._GCP_IDENTITY_REQUEST_URL,
-            params={"audience": "some-audience", "format": "full"},
+            fields={"audience": "some-audience", "format": "full"},
             headers={"Metadata-Flavor": "Google"},
             timeout=30,
         )
@@ -498,27 +493,16 @@ def test_detect_gcp_request_timeout(monkeypatch):
     )
     monkeypatch.setattr(ambient, "_open", lambda fn: stub_file)  # type: ignore
 
-    resp = pretend.stub(raise_for_status=pretend.raiser(Timeout))
-    requests = pretend.stub(
-        get=pretend.call_recorder(lambda url, **kw: resp),
-        HTTPError=HTTPError,
-        Timeout=Timeout,
+    u3 = pretend.stub(
+        request=pretend.raiser(ValueError), exceptions=pretend.stub(MaxRetryError=ValueError)
     )
-    monkeypatch.setattr(ambient, "requests", requests)
+    monkeypatch.setattr(ambient, "urllib3", u3)
 
     with pytest.raises(
         ambient.AmbientCredentialError,
         match=r"GCP: OIDC token request timed out",
     ):
         ambient.detect_gcp("some-audience")
-    assert requests.get.calls == [
-        pretend.call(
-            ambient._GCP_IDENTITY_REQUEST_URL,
-            params={"audience": "some-audience", "format": "full"},
-            headers={"Metadata-Flavor": "Google"},
-            timeout=30,
-        )
-    ]
 
 
 @pytest.mark.parametrize("product_name", ("Google", "Google Compute Engine"))
@@ -533,17 +517,18 @@ def test_detect_gcp(monkeypatch, product_name):
     monkeypatch.setattr(ambient, "logger", logger)
 
     resp = pretend.stub(
-        raise_for_status=lambda: None,
-        text="fakejwt",
+        status=200,
+        data=b"fakejwt",
     )
-    requests = pretend.stub(get=pretend.call_recorder(lambda url, **kw: resp))
-    monkeypatch.setattr(ambient, "requests", requests)
+    u3 = pretend.stub(request=pretend.call_recorder(lambda meth, url, **kw: resp))
+    monkeypatch.setattr(ambient, "urllib3", u3)
 
     assert ambient.detect_gcp("some-audience") == "fakejwt"
-    assert requests.get.calls == [
+    assert u3.request.calls == [
         pretend.call(
+            "GET",
             ambient._GCP_IDENTITY_REQUEST_URL,
-            params={"audience": "some-audience", "format": "full"},
+            fields={"audience": "some-audience", "format": "full"},
             headers={"Metadata-Flavor": "Google"},
             timeout=30,
         )