+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+
+## [1.6.0]
+
 ### Changed
 
 * CircleCI: default to --root-issuer when generating OIDC Token ([[#438](https://github.com/di/id/pull/438)])
@@ -67,7 +70,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Initial split from https://github.com/sigstore/sigstore-python
 
 <!--Release URLs -->
-[Unreleased]: https://github.com/di/id/compare/v1.5.0...HEAD
+[Unreleased]: https://github.com/di/id/compare/v1.6.0...HEAD
+[1.5.0]: https://github.com/di/id/compare/v1.5.0...v1.6.0
 [1.5.0]: https://github.com/di/id/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/di/id/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/di/id/compare/v1.2.1...v1.3.0